last

+++ app/Http/Controllers/ProductController.php
@@ -47,6 +47,7 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        return response()->json('ms');
         $product_data = $request->validate([
             'name'=>['required'],
             'price'=>['required'],
@@ -111,10 +112,8 @@ class ProductController extends Controller
             'image'=>['file' , 'max: 1000'],
             'description'=>['required'],
         ]);
-        if($request->old_image) {
+        if ($request->old_image && $request->file('image')) {
             Storage::delete("/$request->old_image");
-            $product_data['image']=$request->file('image')->store('product_image');
-        } else if ($request->file('image')) {
             $product_data['image']=$request->file('image')->store('product_image');
         }
         Product::where('id', $id)->update($product_data);
